backend: fix migration 0042
 - move `format_financing_method` to `financing_types.py` so that
   the whole migration depepdency tree is not required for running
   this migration
 - fixes `ModuleNotFoundError: No module named 'sqlalchemy'` error when
   running the migration without dev dependencies
